reports-service/requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
weasyprint==63.*
jinja2==3.1.*
matplotlib==3.10.*
apscheduler==3.10.*
sqlalchemy==2.0.*
requests==2.32.*
python-dotenv==1.0.*
python-multipart==0.0.*
